update rename util
