+++ src/connectionStrategies/index.mjs
@@ -1,4 +1,4 @@
-import * as util from '../util/util.js';
+import * as util from '../util/index.mjs';
 
 function abs2rel(value, max) {
