+++ src/dia/Element.mjs
@@ -1,6 +1,6 @@
 import { Cell } from './Cell.mjs';
 import { Point, toRad, normalizeAngle, Rect } from '../g/index.mjs';
-import { isNumber, isObject, interpolate, assign, invoke, normalizeSides } from '../util/util.js';
+import { isNumber, isObject, interpolate, assign, invoke, normalizeSides } from '../util/index.mjs';
 import { elementPortPrototype } from './ports.js';
 
 // Element base model.
