+++ src/dia/ElementView.mjs
@@ -1,4 +1,4 @@
-import { assign, invoke, isFunction, toArray } from '../util/util.js';
+import { assign, invoke, isFunction, toArray } from '../util/index.mjs';
 import { CellView } from './CellView.mjs';
 import { Cell } from './Cell.mjs';
 import V from '../Vectorizer.mjs';
