+++ src/dia/ports.js
@@ -1,4 +1,4 @@
-import * as  util from '../util/util.js';
+import * as  util from '../util/index.mjs';
 import V from '../Vectorizer.mjs';
 import { Rect, Point } from '../g/index.mjs';
 import * as Port from '../layout/ports/port.mjs';
