+++ src/routers/orthogonal.mjs
@@ -1,5 +1,5 @@
 import * as g from '../g/index.mjs';
-import * as util from '../util/util.js';
+import * as util from '../util/index.mjs';
 
 // bearing -> opposite bearing
 var opposites = {
