+++ src/shapes/devs.mjs
@@ -1,6 +1,6 @@
 import { Generic } from './basic.mjs';
 import { Link as LinkBase } from '../dia/Link.mjs';
-import { uniq, difference, assign, toArray, without, isObject } from '../util/util.js';
+import { uniq, difference, assign, toArray, without, isObject } from '../util/index.mjs';
 
 /**
  * @deprecated use the port api instead
