+++ src/util/index.mjs
@@ -1,3 +1,3 @@
 export * from './wrappers.mjs';
-export * from './util.js';
+export * from './util.mjs';
 export * from './cloneCells.mjs';
