--- src/util/util.setTheme.mjs
@@ -1,12 +1,0 @@
-import { invoke } from './util.js';
-import { View, views } from '../mvc/index.mjs';
-
-export const setTheme = function(theme, opt) {
-
-    opt = opt || {};
-
-    invoke(views, 'setTheme', theme, opt);
-
-    // Update the default theme on the view prototype.
-    View.prototype.defaultTheme = theme;
-};
